consolildate flags.ts into types.ts
