+++ src/command/render/latexmk/latexmk.ts
@@ -26,8 +26,8 @@ import {
 import { Format } from "../../../config/types.ts";
 import { pdfEngine } from "../../../config/pdf.ts";
 
-import { PandocOptions, RenderOptions } from "../types.ts";
-import { kStdOut, RenderFlags, replacePandocArg } from "../flags.ts";
+import { PandocOptions, RenderFlags, RenderOptions } from "../types.ts";
+import { kStdOut, replacePandocArg } from "../flags.ts";
 import { OutputRecipe } from "../output.ts";
 import { generatePdf } from "./pdf.ts";
 import { LatexmkOptions } from "./types.ts";
