+++ src/project/types/project-type.ts
@@ -5,8 +5,7 @@
 *
 */
 
-import { PandocFlags } from "../../config/flags.ts";
-import { Format, FormatExtras } from "../../config/types.ts";
+import { Format, FormatExtras, PandocFlags } from "../../config/types.ts";
 import { Metadata } from "../../config/types.ts";
 
 import { PandocRenderer, RenderOptions } from "../../command/render/types.ts";
