+++ src/project/types/website/website-navigation.ts
@@ -27,9 +27,9 @@ import {
   kDependencies,
   kHtmlPostprocessors,
   kSassBundles,
+  PandocFlags,
   SassBundle,
 } from "../../../config/types.ts";
-import { PandocFlags } from "../../../config/flags.ts";
 import {
   hasTableOfContents,
   hasTableOfContentsTitle,
