+++ src/project/types/website/website.ts
@@ -31,9 +31,9 @@ import {
   FormatExtras,
   kDependencies,
   kHtmlPostprocessors,
+  PandocFlags,
 } from "../../../config/types.ts";
 import { isHtmlOutput } from "../../../config/format.ts";
-import { PandocFlags } from "../../../config/flags.ts";
 
 import {
   kIncludeInHeader,
